Add Support for Jobs Ending in "UNKNOWN" State
Adding support for jobs that end in an "UKNOWN" state. Jobs in unknown states are treated as a failure state instead of causing a panic. If possible the sequence should be retried as with a normal failure and if that is not possible or it is not able to be retried then the request should end in failure and not a panic.